Refactor of chat_client.cpp is complete.
Actions executed across six subtasks:
1. Extracted model-related logic → chat_client_models.[h|cpp] (Step 1).
2. Extracted CURL/API helpers → chat_client_api.[h|cpp] (Step 2).
3. Extracted interactive turn workflow → chat_client_turn.[h|cpp] (Step 3).
4. Centralised shared helpers → chat_client_internal.hpp (Step 4).
5. Performed header-hygiene & forward-declarations across all new headers (Step 5).
6. Consolidated CMakeLists.txt: grouped the four chat-client translation units under one PRIVATE target_sources block and guarded an MSVC source_group (Step 6).

Build & quality gates:
• Full build passes with -Wall -Wextra -pedantic, zero new warnings.
• All 7/7 unit tests green.
• scripts/loc_report.py (71 files scanned) – every project-owned file ≤ 500 LOC; new max 437 LOC.

chat_client.cpp now 122 LOC; refactor met the roadmap objective and introduces no behavioural changes.
